Alphabet submodule refactor.
